+++ packages/slider/stories/slider.stories.ts
@@ -499,6 +499,22 @@ hideStepper.decorators = [editableDecorator];
 
 export const Gradient = (args: StoryArgs = {}): TemplateResult => {
     return html`
+        <style>
+            sp-slider {
+                --mod-slider-track-color: linear-gradient(
+                    to right,
+                    red,
+                    green 100%
+                );
+            }
+            sp-slider[dir='rtl'] {
+                --mod-slider-track-color: linear-gradient(
+                    to left,
+                    red,
+                    green 100%
+                );
+            }
+        </style>
         <div
             style="
                 width: 500px;
@@ -506,10 +522,6 @@ export const Gradient = (args: StoryArgs = {}): TemplateResult => {
             "
         >
             <sp-slider
-                style="
-                    --spectrum-slider-track-color:linear-gradient(to right, red, green 100%);
-                    --spectrum-slider-track-color-rtl:linear-gradient(to left, red, green 100%);
-                "
                 label="Opacity"
                 max="100"
                 min="0"