+++ projects/documentation/src/components.ts
@@ -11,6 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import './router.js';
+import './components/demo-container.js';
 import '@spectrum-web-components/tabs/sp-tab-panel.js';
 import '@spectrum-web-components/tabs/sp-tab.js';
 import '@spectrum-web-components/tabs/sp-tabs.js';
@@ -44,6 +45,10 @@ import '@spectrum-web-components/table/sp-table-checkbox-cell.js';
 import '@spectrum-web-components/table/sp-table-head.js';
 import '@spectrum-web-components/table/sp-table-head-cell.js';
 import '@spectrum-web-components/table/sp-table-row.js';
+import '@spectrum-web-components/field-group/sp-field-group.js';
+import '@spectrum-web-components/field-label/sp-field-label.js';
+import '@spectrum-web-components/card/sp-card.js';
+import '@spectrum-web-components/icon/sp-icon.js';
 
 if ('requestIdleCallback' in window) {
     requestIdleCallback(() => {